refactor: 하나의 view에서 InputView,OutputView로 분리[A-3]
