+++ src/App.js
@@ -1,5 +1,4 @@
 import BaseBallController from './controller/index.js';
-import View from './view/index.js';
 
 class App {
   #controller;
@@ -9,9 +8,11 @@ class App {
   }
 
   async play() {
-    View.printStart();
     await this.#controller.run();
   }
 }
 
 export default App;
+
+const app = new App();
+app.play();
