+++ src/view/OutputView.js
@@ -1,4 +1,5 @@
 import { Console } from '@woowacourse/mission-utils';
+import { HINT, OUTPUT_MESSAGES } from '../constants/Messages.js';
 
 const OutputView = {
   /**
@@ -7,6 +8,24 @@ const OutputView = {
    */
   print(message) {
     Console.print(message);
+  },
+
+  printStart() {
+    OutputView.print(OUTPUT_MESSAGES.game_start);
+  },
+
+  printSuccess() {
+    OutputView.print(OUTPUT_MESSAGES.game_success);
+  },
+
+  printHint({ ball, strike }) {
+    if (ball === 0 && strike === 0) {
+      OutputView.print(HINT.nothing);
+    } else {
+      const message = HINT.message(ball, strike);
+      const hint = message.replace(HINT.clear, '').trim();
+      OutputView.print(hint);
+    }
   },
 };
